+++ CHANGELOG.md
@@ -15,12 +15,13 @@ We go to the next version after each release on
 
 ### Fixed
 
-- `.set_icon_size` is not a function error (Date Menu).
 - a11y error in GNOME Shell mobile.
 - Activities button error in GNOME Shell mobile.
 - Marked as urgent signal for window demand attention focus feature.
 - Panel flickering when its position is at the bottom.
+- `.set_icon_size` is not a function error (Date Menu).
 - Unwanted padding on disabled power icon.
+- Unwanted shadow for workspace in app grid.
 
 ### Added